+++ braindecode/models/util.py
@@ -217,4 +217,5 @@ models_mandatory_parameters = [
     ("USleep", ["n_chans", "n_outputs", "n_times", "sfreq"], dict(sfreq=128)),
     ("BIOT", ["n_chans", "n_outputs", "sfreq"], None),
     ("Labram", ["n_chans", "n_outputs", "n_times"], None),
+    ("EEGSimpleConv", ["n_chans", "n_outputs", "sfreq"], None),
 ]