+++ docs/whats_new.rst
@@ -23,11 +23,11 @@ Current 0.9 (dev0)
 
 Enhancements
 ~~~~~~~~~~~~
-- Adding :class:`braindecode.datasets.NMT` dataset (:gh:`443` by `MohammadJavadD`_)
-
-- Add an integration test for all the models (:gh: 570 by `Bruno Aristimunha`_)
+- Adding :class:`braindecode.datasets.NMT` dataset (:gh:`443` by `Mohammad Javad D`_  and `Bruno Aristimunha`_)
+- Adding an integration test for all the models (:gh: 570 by `Bruno Aristimunha`_)
 - Adding :class:`braindecode.models.BIOT` (:gh:`573` by `Bruno Aristimunha`_)
 - Adding :class:`braindecode.models.Labram` (:gh:`578` by `Bruno Aristimunha`_)
+- Adding :class:`braindecode.models.EEGSimpleConv` (:gh:`581` by `Yassine El Ouahidi`_ and `Bruno Aristimunha`_)
 
 
 Bugs
@@ -251,3 +251,4 @@ Authors
 .. _Sylvain Chevallier: https://github.com/sylvchev
 .. _Remi Delbouys: https://github.com/remidbs
 .. _Valentin Iovene: https://github.com/tgy
+.. _Yassine El Ouahidi: https://github.com/elouayas